# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ulimax — Sistema de Controle de Projetos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /automation-rules:
    get:
      operationId: listAutomationRules
      tags: [automation]
      summary: List automation rules
      responses:
        "200":
          description: List of automation rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationRule"
    post:
      operationId: createAutomationRule
      tags: [automation]
      summary: Create an automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationRuleInput"
      responses:
        "201":
          description: Created automation rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
  /automation-rules/{id}:
    patch:
      operationId: updateAutomationRule
      tags: [automation]
      summary: Update an automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationRuleInput"
      responses:
        "200":
          description: Updated automation rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
    delete:
      operationId: deleteAutomationRule
      tags: [automation]
      summary: Delete an automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /invitations:
    get:
      operationId: listInvitations
      tags: [invitations]
      summary: List pending invitations (gestor only)
      responses:
        "200":
          description: List of pending invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invite"
        "403":
          description: Forbidden
    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Send a team invitation (gestor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteBody"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "403":
          description: Forbidden
        "409":
          description: Already invited or already has an account
  /invitations/{id}:
    delete:
      operationId: deleteInvitation
      tags: [invitations]
      summary: Revoke a pending invitation (gestor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Invitation revoked
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List project templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectTemplate"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a project template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectTemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectTemplate"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template with its tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template with tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectTemplateFull"
        "404":
          description: Not found
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /templates/{id}/tasks:
    post:
      operationId: addTemplateTask
      tags: [templates]
      summary: Add a task to a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectTemplateTaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectTemplateTask"
  /templates/{id}/tasks/{taskId}:
    delete:
      operationId: deleteTemplateTask
      tags: [templates]
      summary: Remove a task from a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /templates/{id}/apply:
    post:
      operationId: applyTemplate
      tags: [templates]
      summary: Create a project from a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyTemplateInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
components:
  schemas:
...
    ProjectParticipantSummary:
      type: object
      required: [memberId, memberName]
      properties:
        memberId:
          type: integer
        memberName:
          type: string
        memberAvatarUrl:
          type: ["string", "null"]
    Project:
      type: object
      required: [id, name, status, priority, createdAt, participants, taskTotal, taskDone]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [a_iniciar, em_projeto, em_aprovacao, em_producao, aguardando_instalacao, em_instalacao]
        priority:
          type: string
          enum: [low, medium, high]
        taskTotal:
          type: integer
        taskDone:
          type: integer
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        finalDate:
          type: ["string", "null"]
        producaoStartDate:
          type: ["string", "null"]
        producaoEndDate:
          type: ["string", "null"]
        producaoFinalDate:
          type: ["string", "null"]
        medicaoDate:
          type: ["string", "null"]
        instalacaoStartDate:
          type: ["string", "null"]
        materialType:
          type: ["string", "null"]
          enum: [madeira, aluminio]
        approvalStatus:
          type: ["string", "null"]
          enum: [pending, approved, rejected]
        approvalNote:
          type: ["string", "null"]
        approvalAt:
          type: ["string", "null"]
        approvalBy:
          type: ["integer", "null"]
        createdAt:
          type: string
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ProjectParticipantSummary"
...
    UserRole:
      type: string
      enum: [gestor, executor, observador, gestor_obras]
...
    Invite:
      type: object
      required: [id, email, name, intendedRole, invitedAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        intendedRole:
          $ref: "#/components/schemas/UserRole"
        clerkInvitationId:
          type: ["string", "null"]
        invitedAt:
          type: string
        emailSent:
          type: boolean
        signUpUrl:
          type: string
    CreateInviteBody:
      type: object
      required: [email, name, intendedRole]
      properties:
        email:
          type: string
        name:
          type: string
        intendedRole:
          $ref: "#/components/schemas/UserRole"
...
    ProjectTemplate:
      type: object
      required: [id, name, priority, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high]
        taskCount:
          type: integer
        createdAt:
          type: string
    ProjectTemplateFull:
      type: object
      required: [id, name, priority, createdAt, tasks]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/ProjectTemplateTask"
    ProjectTemplateTask:
      type: object
      required: [id, templateId, title, priority, offsetDays]
      properties:
        id:
          type: integer
        templateId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high]
        offsetDays:
          type: integer
    ProjectTemplateInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
    ProjectTemplateTaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        offsetDays:
          type: integer
    ApplyTemplateInput:
      type: object
      required: [name, startDate]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        startDate:
          type: string
...
    AutomationRule:
      type: object
      required: [id, name, trigger, actionType, isActive, executionCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        trigger:
          type: string
          enum: [task_completed, task_status_changed, project_completed, task_assigned, project_status_changed]
        actionType:
          type: string
          enum: [notify_assignee, notify_all, notify_gestor, advance_task_status]
        isActive:
          type: boolean
        executionCount:
          type: integer
        lastFiredAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    AutomationRuleInput:
      type: object
      required: [name, trigger, actionType]
      properties:
        name:
          type: string
          minLength: 1
        trigger:
          type: string
          enum: [task_completed, task_status_changed, project_completed, task_assigned, project_status_changed]
        actionType:
          type: string
          enum: [notify_assignee, notify_all, notify_gestor, advance_task_status]
        isActive:
          type: boolean
...