# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ulimax — Sistema de Controle de Projetos
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  gestor: 'gestor',
  executor: 'executor',
  observador: 'observador',
  gestor_obras: 'gestor_obras',
} as const;